migrate vote endpoint
