+++ .bruno/vote.bru
@@ -1,0 +1,21 @@
+meta {
+  name: vote
+  type: http
+  seq: 3
+}
+
+post {
+  url: {{api}}/vote
+  body: json
+  auth: none
+}
+
+headers {
+  Authorization: tekohhuge12369
+}
+
+body:json {
+  {
+    "user": "672793821850894347"
+  }
+}
